datas

+++ src/main/kotlin/hu/zsof/restaurantApp/model/Category.kt
@@ -8,8 +8,7 @@ data class Category(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false, updatable = false)
-    val id: Int,
-
+    val id: Long,
     val name: String? = null,
 
   /*  @OneToMany(mappedBy = "category", cascade = [CascadeType.ALL])
